backend/requirements.txt
openai-whisper
torch
torchaudio
langchain
langchain-core
langchain-text-splitters
langchain-chroma
langchain-ollama
pandas
pyttsx3
sounddevice
numpy
fastapi
uvicorn